Updated readme with full docker-compose example
Thx for building this great tool!

As I'm spawning my containers with docker-compose I just created a full `docker-compose.yaml` based on the parts you mentioned in this readme.  I think having a full example (and not only parts) might also be useful to others. So this patch adds a full file for copy-and-pasting.