Refined development environment setup
Streamlined the devcontainer configuration for a more focused Node.js development experience. This involved simplifying the container by removing Postgres integration and many customized settings not universally applicable, such as those specific for NextJS development and various unnecessary VS Code settings and extensions. Also, cleaned up the Zsh plugins list to focus on essentials, improving setup time and reducing potential sources of confusion or conflict for new developers. This change aims to make the development environment quicker to set up, easier to understand, and more broadly applicable.

Removed settings and plugins are now considered out of scope for the primary development goals but can still be added on a per-project basis as needed.