Restoring focus outline on links
Browsers have a focus outline for a reason: accessibility, letting users interact with the page using just the keyboard.

We, web developers, must not remove the focus outline without providing an equivalent alternative.

* https://accessiblewebsiteservices.com/accessible-css-focus-on-the-outline/
* https://theadminbar.com/accessibility-weekly/focus-outlines/
* https://stackoverflow.com/questions/9274948/css-outline-best-practices

I considered using the same visual feedback as the mouse-over effect, but that's not possible in the current HTML/CSS. That `transform: translateY(-3px)` effect is being applied to the parent `div.card`, which is not focusable, and thus we cannot trigger the same effect. The solution would require reworking both the HTML and CSS. (It can also be argued that tiny 3px offset is too subtle to be used as a focus indicator.)

For now, I've restored the default browser focus outline, and someday in the future someone could try to provide a different focus indication.